renamed internal/xsql/propose to internal/xsql/xquery
