+++ internal/xsql/xquery/rows.go
@@ -1,4 +1,4 @@
-package propose
+package xquery
 
 import (
 	"context"
